backend/requirements.txt
Flask
flask-cors
python-dotenv
twilio
openai
requests
supabase
gunicorn
python-dateutil
pytz
redis
celery
pydantic